+++ taxcalc/policy.py
@@ -8,6 +8,7 @@ Tax-Calculator federal tax policy Policy class.
 import os
 import json
 from pathlib import Path
+import paramtools
 from taxcalc.parameters import Parameters
 from taxcalc.growfactors import GrowFactors
 
@@ -62,15 +63,6 @@ class Policy(Parameters):
         'DependentCredit_before_CTC': 'is a removed parameter name',
         'FilerCredit_c': 'is a removed parameter name',
         'ALD_InvInc_ec_base_RyanBrady': 'is a removed parameter name',
-        # following parameter renamed in PR 2292 merged on 2019-04-15
-        "cpi_offset": (
-            "was renamed parameter_indexing_CPI_offset. "
-            "See documentation for change in usage."
-        ),
-        "CPI_offset": (
-            "was renamed parameter_indexing_CPI_offset. "
-            "See documentation for change in usage."
-        ),
         # following parameters renamed in PR 2345 merged on 2019-06-24
         'PT_excl_rt':
         'was renamed PT_qbid_rt in release 2.4.0',
@@ -140,13 +132,34 @@ class Policy(Parameters):
         """
         return Parameters._read_json_revision(obj, 'policy')
 
-    def implement_reform(self, reform,
+    def implement_reform(self, reform: dict,
                          print_warnings=True, raise_errors=True):
         """
-        Implement reform using Tax-Calculator syled reforms/adjustments. Users
-        may also use the adjust method with ParamTools styled reforms.
+        Implement reform using a Tax-Calculator-style reform dictionary.
         """
-        # need to do conversion:
+        if not isinstance(reform, dict):
+            raise paramtools.ValidationError(
+                {'errors': {'schema': 'reform must be a dictionary'}},
+                None
+            )
+        deprecated_parameters = [
+            'ID_AmountCap_Switch',
+            'ID_AmountCap_rt',
+            'ID_BenefitCap_Switch',
+            'ID_BenefitCap_rt',
+            'ID_BenefitSurtax_Switch',
+            'ID_BenefitSurtax_crt',
+            'ID_BenefitSurtax_trt',
+            'ID_BenefitSurtax_em',
+        ]
+        for param in reform.keys():
+            if param in deprecated_parameters:
+                print(
+                    f'DEPRECATION WARNING: the {param} policy parameter\n'
+                    'is scheduled to be removed in Tax-Calculator 5.0.0;\n'
+                    'if you think this removal should not happen, open an\n'
+                    'issue on GitHub to make your case for non-removal.'
+                )
         return self._update(reform, print_warnings, raise_errors)
 
     @staticmethod
@@ -170,9 +183,9 @@ class Policy(Parameters):
             vo["value"] for
             vo in self._data["parameter_indexing_CPI_offset"]["value"]
         ]
-        # policy_current_law.json should not specify any non-zero values
-        # for the parameter_indexing_CPI_offset parameter, so check this
-        assert any(cpi_vals) is False
+        # parameter_indexing_CPI_offset parameter is no longer used, so check:
+        assert any(cpi_vals) is False, \
+            'obsolete parameter_indexing_CPI_offset values must all be zero'
         syr = max(self.start_year, self._gfactors.first_year)
         self._inflation_rates = self._gfactors.price_inflation_rates(
             syr, self.end_year